src/A3/requirements.txt
# A3 before / after 共用 Python 依赖
# 安装方式：
#   python3 -m pip install -r src/A3/requirements.txt
# 不要求先升级 pip。
torch>=2.1.0
torchvision>=0.16.0
timm>=1.0.7
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
tabulate>=0.9.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.66.0
safetensors>=0.4.0

# WSI 读取依赖
openslide-python>=1.3.1
tiffslide>=2.5.1
# 说明：
# 1. openslide-python 还需要系统级 OpenSlide 共享库。
# 2. 若目标机不便安装 OpenSlide，可直接拷贝仓库根目录 `_vendor/`，
#    当前代码会在 openslide 不可用时自动尝试 tiffslide 回退。

# ONNX 回退链 / 对齐脚本
onnx>=1.14.0
onnxruntime>=1.17.0

# 昇腾环境说明
# 1. before（PyTorch/NPU）需要与目标机 CANN / PyTorch 匹配的 torch_npu。
#    该包强依赖目标环境，通常优先使用昇腾镜像内预装版本，
#    或手动安装与目标机完全匹配的 wheel，不建议在这里写死版本。
# 2. after（OM/ACL）需要目标机已正确安装 CANN 运行时，
#    并保证 `python3 -c "import acl"` 能成功。
